[IMP] devtools: add __owl__ access.
To be able to make devtools, we need access to the __owl__ variable
on a component. It contains most of the information a devtool may
required.

Just like in Vue js, we attach it to the HTMLElement of a component.

For its update, we use the onMounted and onPatched hook.